Digital_Hackers_Source_Code/Digital_Hackers_Source_Code/backend/restaurant-reservation-system/src/api/app.py
from fastapi import FastAPI
from api.routes import reservations, tables, guests

app = FastAPI()

app.include_router(reservations.router)
app.include_router(tables.router)
app.include_router(guests.router)