Digital_Hackers_Source_Code/Digital_Hackers_Source_Code/backend/restaurant-reservation-system/requirements.txt
boto3
botocore
awscli
opensearch-py
requests-aws4auth
pyyaml
retrying
pandas
strands-agents>=0.2.1
strands-agents-tools>=1.0.0
python-dotenv==1.0.0
pydantic[email]
pytest